% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.scaleL}
\alias{print.scaleL}
\title{Print a \code{scaleL} fit}
\usage{
\method{print}{scaleL}(x, ...)
}
\arguments{
\item{x}{A \code{scaleL} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a \code{scaleL} fit
}
