% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.scaleL}
\alias{summary.scaleL}
\alias{print.summary.scaleL}
\title{Summarize a \code{scaleL} fit}
\usage{
\method{summary}{scaleL}(object, ...)

\method{print}{summary.scaleL}(x, ...)
}
\arguments{
\item{object}{A \code{scaleL} object.}

\item{...}{Ignored.}

\item{x}{A \code{summary.scaleL} object.}
}
\value{
A \code{summary.scaleL} object.
}
\description{
Summarize a \code{scaleL} fit
}
