% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\docType{data}
\name{TIER1_SHIFTED_PRIOR}
\alias{TIER1_SHIFTED_PRIOR}
\title{Tier 1 shifted prior (sensitivity arm)}
\format{
Named numeric vector.
}
\usage{
TIER1_SHIFTED_PRIOR
}
\description{
Variant of \link{TIER1_PRIOR} with more mass on L = 5 and L = 7, for use as
a sensitivity-analysis arm.
}
\keyword{datasets}
