% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleL-package.R
\docType{package}
\name{scaleL-package}
\alias{scaleL-package}
\title{scaleL: Bayesian Imputation of Likert Scale Length for Meta-Analysis}
\description{
Bayesian imputation of missing Likert scale length L in meta-analytic
data sets, with calibrated uncertainty propagation via multiple
imputation and Rubin's rules.
}
\details{
The main user-facing function is \code{\link[=scaleL]{scaleL()}}. See the package README for
usage examples and scope-of-applicability notes.
}
\author{
\strong{Maintainer}: scaleL Authors \email{scaleL.pkg@proton.me}

}
\keyword{internal}
